+++ packages/shared-types/src/cheeky_pony_shared/openapi-types.ts
@@ -44,6 +44,8 @@ export interface components {
     "LoginRequest": unknown;
     "LoginResponse": unknown;
     "OuiLookupResponse": unknown;
+    "ReadinessCheck": unknown;
+    "ReadinessCheckId": unknown;
     "RegisterRequest": unknown;
     "ReportCreateRequest": unknown;
     "ReportCreateResponse": unknown;